chore: rename personal_data to schema_rb
