+++ pkg/detectors/rails/schema_rb/schema_rb.go
@@ -1,4 +1,4 @@
-package personal_data
+package schema_rb
 
 import (
 	"strings"
